Split application-package
Fixes #1356
