+++ dev-packages/application-manager/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@theia/application-package",
+  "name": "@theia/application-manager",
   "version": "0.3.7",
-  "description": "Theia application package API.",
+  "description": "Theia application manager API.",
   "publishConfig": {
     "access": "public"
   },
@@ -29,15 +29,22 @@
     "docs": "theiaext docs"
   },
   "dependencies": {
-    "@types/fs-extra": "^4.0.2",
-    "@types/request": "^2.0.3",
-    "@types/semver": "^5.4.0",
-    "@types/write-json-file": "^2.2.1",
-    "changes-stream": "^2.2.0",
-    "fs-extra": "^4.0.2",
-    "request": "^2.82.0",
-    "semver": "^5.4.1",
-    "write-json-file": "^2.2.0"
+    "@theia/application-package": "0.3.7",
+    "bunyan": "^1.8.10",
+    "circular-dependency-plugin": "^2.0.0",
+    "copy-webpack-plugin": "^4.0.1",
+    "css-loader": "^0.28.1",
+    "electron": "1.7.11",
+    "electron-rebuild": "^1.5.11",
+    "file-loader": "^0.11.1",
+    "font-awesome-webpack": "0.0.5-beta.2",
+    "ignore-loader": "^0.1.2",
+    "less": "^2.7.2",
+    "source-map-loader": "^0.2.1",
+    "source-map-support": "^0.4.18",
+    "umd-compat-loader": "^2.1.1",
+    "url-loader": "^0.5.8",
+    "webpack": "^2.2.1"
   },
   "devDependencies": {
     "@theia/ext-scripts": "^0.2.0"
