+++ dev-packages/application-manager/src/generator/frontend-generator.ts
