+++ dev-packages/application-manager/src/generator/index.ts
