+++ dev-packages/application-manager/src/generator/webpack-generator.ts
