+++ dev-packages/application-manager/src/index.ts
@@ -1,10 +1,10 @@
 /*
- * Copyright (C) 2017 TypeFox and others.
+ * Copyright (C) 2018 Ericsson and others.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-export * from './npm-registry';
-export * from './extension-package';
-export * from './application-package';
+export * from './rebuild';
+export * from './application-package-manager';
+export * from './application-process';
