+++ dev-packages/application-manager/src/rebuild.ts
