+++ packages/extension-manager/package.json
@@ -3,7 +3,7 @@
   "version": "0.3.7",
   "description": "Theia - Extension Manager",
   "dependencies": {
-    "@theia/application-package": "^0.3.7",
+    "@theia/application-manager": "^0.3.7",
     "@theia/core": "^0.3.7",
     "@theia/filesystem": "^0.3.7",
     "@types/fs-extra": "^4.0.2",
